fixes #9916 - initialise Puppet using public APIs
Previously Puppet was never initialised, so its internal settings weren't ready
for use.  This causes an error under 3.7.5 where it expects the internal
'context' to be set up to retrieve environment-specific parser config, but it
hasn't been.

This code is the same as the smart proxy's Proxy::Puppet::Initializer, with the
exception of defining a configuration file (default should suffice here).