Move puppet dev dep to Gemfile, add puppet-strings dep
The 'PUPPET_VERSION' environment variable will be used to install
specific versions of Puppet, and if the version specified supports
puppet-strings, then the dep's added to test the strings parser.

This allows the removal of Gemfile alterations from the Jenkins job and
makes it easier to test different versions in development.